third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-css-in-iframe-expected.txt

CONSOLE WARNING: line 4: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-css.html' was loaded over HTTPS, but ran insecure content from 'http://127.0.0.1:8080/security/mixedContent/resources/style.css': this content should also be loaded over HTTPS.

This test loads a secure iframe that loads an insecure style sheet. We should trigger a mixed content callback because an active network attacker can use CSS3 to breach the confidentiality of the HTTPS security origin.
